Finished the octree (reasonably)
It isnt perfect as it only seems to be able to handle about 50 entities
before slowing down significantly. Possibilities for fixing include
modifying removal of nodes not at rest, and simply building and rebuilding
the octree every frame... Also need to consider putting the octree nodes
onto the stack, as I cant currently see a reason for them being on the
heap!