GJK Almost done
Collision detection seems to work in the majority of cases. Might need to
go and find the case that doesnt work... But that shouldn't be too much of
an issue!!! Now to organise the new detection properly, and the get some
collision response data... I dont think proper data is that necessary, so
will try using spheres first. This could look terrible, but remember, its
about making it look believable!!